python/requirements.txt
pandas
mysql-connector-python
prophet
numpy<2.0
plotly